# docs/openapi/custody-v1.yaml
openapi: 3.1.0
info:
  title: Surprising Wallet Custody API
  version: 1.0.0
  description: |
    Multi-tenant custody, Console, and platform control-plane API.
    Tenant identity is derived from the authenticated credential and is never
    accepted as a trusted request field.
servers:
  - url: https://wallet.example.com
...
paths:
...
  /custody/console/v1/webhooks:
    get:
      tags: [Console Webhooks]
      summary: List tenant Webhook endpoints
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Endpoints and 24-hour health }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Console Webhooks]
      summary: Create a Webhook endpoint
      description: The signing secret is returned only once.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWebhook' }
      responses:
        '200': { description: Pending endpoint and one-time signing secret }
        default: { $ref: '#/components/responses/Error' }
  /custody/console/v1/webhooks/{endpointId}/verify:
    post:
      tags: [Console Webhooks]
      summary: Send a challenge and activate an endpoint
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/EndpointId'
      responses:
        '200': { description: Verified endpoint }
        default: { $ref: '#/components/responses/Error' }
  /custody/console/v1/webhooks/{endpointId}/status:
    patch:
      tags: [Console Webhooks]
      summary: Enable or disable a verified endpoint
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/EndpointId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Error' }
...
  /custody/platform/v1/auth/login:
    post:
      tags: [Platform]
      summary: Sign in as platform administrator
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Platform session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResult' }
        default: { $ref: '#/components/responses/Error' }
  /custody/platform/v1/auth/me:
    get:
      tags: [Platform]
      summary: Read the platform principal
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Current principal }
        default: { $ref: '#/components/responses/Error' }
  /custody/platform/v1/auth/logout:
    post:
      tags: [Platform]
      summary: Revoke the platform session
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Error' }
  /custody/platform/v1/tenants:
    get:
      tags: [Platform]
      summary: List tenants and activity counts
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200': { description: Tenant records }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Platform]
      summary: Create a tenant and initial tenant administrator
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, name, adminEmail, adminDisplayName, adminPassword]
              properties:
                slug:
                  type: string
                  pattern: '^[a-z0-9][a-z0-9-]{1,62}[a-z0-9]$'
                name: { type: string, maxLength: 160 }
                adminEmail: { type: string, format: email }
                adminDisplayName: { type: string, maxLength: 120 }
                adminPassword: { type: string, minLength: 12, maxLength: 256 }
      responses:
        '200': { description: Created tenant }
        default: { $ref: '#/components/responses/Error' }
  /custody/platform/v1/tenants/{tenantId}:
    get:
      tags: [Platform]
      summary: Read a tenant
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200': { description: Tenant record }
        default: { $ref: '#/components/responses/Error' }
  /custody/platform/v1/tenants/{tenantId}/status:
    patch:
      tags: [Platform]
      summary: Suspend or reactivate a tenant
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [ACTIVE, SUSPENDED] }
      responses:
        '200': { description: Updated tenant }
        default: { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ConsoleSession
    custodyKey:
      type: apiKey
      in: header
      name: X-Custody-Key
    custodyTimestamp:
      type: apiKey
      in: header
      name: X-Custody-Timestamp
    custodyNonce:
      type: apiKey
      in: header
      name: X-Custody-Nonce
    custodySignature:
      type: apiKey
      in: header
      name: X-Custody-Signature
  parameters:
...
    TenantId:
      name: tenantId
      in: path
      required: true
      schema: { type: string, format: uuid }
    EndpointId:
      name: endpointId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
  responses:
    Ok:
      description: Operation completed
      content:
        application/json:
          schema:
            type: object
            required: [ok]
            properties:
              ok: { type: boolean, const: true }
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
    LoginResult:
      type: object
      required: [token, expiresAt, userId, email, displayName, role]
      properties:
        token: { type: string }
        expiresAt: { type: string, format: date-time }
        userId: { type: string, format: uuid }
        tenantId: { type: [string, 'null'], format: uuid }
        tenantSlug: { type: [string, 'null'] }
        email: { type: string, format: email }
        displayName: { type: string }
        role: { type: string }
...
    CreateWebhook:
      type: object
      required: [name, url, events]
      properties:
        name: { type: string, maxLength: 120 }
        url: { type: string, format: uri, maxLength: 2048 }
        events:
          type: array
          uniqueItems: true
          items:
            type: string
            enum:
              - DEPOSIT.CONFIRMED
              - WITHDRAWAL.CREATED
              - WITHDRAWAL.BROADCAST
              - WITHDRAWAL.BROADCAST_UNKNOWN
              - WITHDRAWAL.CONFIRMED
              - WITHDRAWAL.FAILED